+++ src/main/java/com/travelland/global/config/WebSecurityConfig.java
@@ -3,8 +3,8 @@ package com.travelland.global.config;
 import com.travelland.global.jwt.JwtUtil;
 import com.travelland.global.security.JwtAuthorizationFilter;
 import com.travelland.global.security.UserDetailsServiceImpl;
-import com.travelland.repository.MemberRepository;
-import com.travelland.repository.RefreshTokenRepository;
+import com.travelland.repository.member.MemberRepository;
+import com.travelland.repository.member.RefreshTokenRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
