+++ src/main/java/com/travelland/global/security/JwtAuthorizationFilter.java
@@ -5,8 +5,8 @@ import com.travelland.domain.RefreshToken;
 import com.travelland.global.exception.CustomException;
 import com.travelland.global.exception.ErrorCode;
 import com.travelland.global.jwt.JwtUtil;
-import com.travelland.repository.MemberRepository;
-import com.travelland.repository.RefreshTokenRepository;
+import com.travelland.repository.member.MemberRepository;
+import com.travelland.repository.member.RefreshTokenRepository;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
