+++ src/main/java/com/travelland/repository/member/RefreshTokenRepository.java
@@ -1,4 +1,4 @@
-package com.travelland.repository;
+package com.travelland.repository.member;
 
 import com.travelland.domain.RefreshToken;
 import org.springframework.data.repository.CrudRepository;
