+++ src/main/java/com/travelland/service/KakaoService.java
@@ -9,8 +9,8 @@ import com.travelland.domain.Member;
 import com.travelland.domain.RefreshToken;
 import com.travelland.dto.MemberDto;
 import com.travelland.global.jwt.JwtUtil;
-import com.travelland.repository.MemberRepository;
-import com.travelland.repository.RefreshTokenRepository;
+import com.travelland.repository.member.MemberRepository;
+import com.travelland.repository.member.RefreshTokenRepository;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
