+++ src/main/java/com/travelland/service/plan/PlanScrapService.java
@@ -6,7 +6,7 @@ import com.travelland.domain.plan.PlanScrap;
 import com.travelland.dto.PlanDto;
 import com.travelland.global.exception.CustomException;
 import com.travelland.global.exception.ErrorCode;
-import com.travelland.repository.MemberRepository;
+import com.travelland.repository.member.MemberRepository;
 import com.travelland.repository.plan.PlanRepository;
 import com.travelland.repository.plan.PlanScrapRepository;
 import lombok.RequiredArgsConstructor;
