+++ src/main/java/com/travelland/service/trip/TripScrapService.java
@@ -6,7 +6,7 @@ import com.travelland.domain.TripScrap;
 import com.travelland.dto.TripDto;
 import com.travelland.global.exception.CustomException;
 import com.travelland.global.exception.ErrorCode;
-import com.travelland.repository.MemberRepository;
+import com.travelland.repository.member.MemberRepository;
 import com.travelland.repository.trip.TripRepository;
 import com.travelland.repository.trip.TripScrapRepository;
 import lombok.RequiredArgsConstructor;
